Sort out directory length calculation.
The directory entry that has the accurate directory size in it is the
one in the first 128 bytes of the directory cluster itself (i.e. the
entry for '.'). When opening a FileHandle for a directory entry that's
inaccurate, mark the FileHandle and fix it up when the first read is
done.